+++ bundles/org.openhab.ui/web/src/components/widgets/modals/oh-sheet.vue
@@ -5,7 +5,9 @@
       <div class="right"><f7-link sheet-close>Close</f7-link></div>
     </f7-toolbar>
 
-    <component :is="componentType" :context="context" :class="{notready: !ready}" />
+    <component v-if="visibleToCurrentUser" :is="componentType" :context="context" :class="{notready: !ready}" />
+    <empty-state-placeholder v-if="page && !visibleToCurrentUser" icon="multiply_circle_fill" title="page.unavailable.title" text="page.unavailable.text" />
+
   </f7-sheet>
 </template>
 
@@ -15,48 +17,9 @@
 </style>
 
 <script>
-import OhLayoutPage from '@/components/widgets/layout/oh-layout-page.vue'
+import modal from './modal-mixin'
 
 export default {
-  components: {
-    'oh-layout-page': OhLayoutPage,
-    'oh-map-page': () => import('@/components/widgets/map/oh-map-page.vue'),
-    'oh-plan-page': () => import('@/components/widgets/plan/oh-plan-page.vue'),
-    'oh-chart-page': () => import('@/components/widgets/chart/oh-chart-page.vue')
-  },
-  props: ['uid', 'modalParams'],
-  data () {
-    return {
-    }
-  },
-  computed: {
-    context () {
-      return {
-        component: Object.assign({}, this.page || this.widget || this.standard, { config: this.modalParams }),
-        store: this.$store.getters.trackedItems,
-        config: this.modalParams
-      }
-    },
-    page () {
-      return (this.uid.indexOf('page:') === 0) ? this.$store.getters.page(this.uid.substring(5)) : null
-    },
-    widget () {
-      return (this.uid.indexOf('widget:') === 0) ? this.$store.getters.widget(this.uid.substring(7)) : null
-    },
-    standard () {
-      return (this.uid.indexOf('oh-') === 0) ? { component: this.uid } : null
-    },
-    ready () {
-      return this.page || this.widget || this.standard
-    },
-    componentType () {
-      if (this.page) {
-        return this.page.component
-      } else if (this.widget || this.standard) {
-        return 'generic-widget-component'
-      }
-      return null
-    }
-  }
+  mixins: [modal]
 }
 </script>
