+++ bundles/org.openhab.ui/web/src/pages/page/page-view.vue
@@ -10,17 +10,19 @@
       </f7-nav-right>
     </f7-navbar>
 
-    <f7-toolbar tabbar labels bottom v-if="page && pageType === 'tabs'">
+    <f7-toolbar tabbar labels bottom v-if="page && pageType === 'tabs' && visibleToCurrentUser">
       <f7-link v-for="(tab, idx) in page.slots.default" :key="idx" tab-link @click="currentTab = idx" :tab-link-active="currentTab === idx" :icon-ios="tab.config.icon" :icon-md="tab.config.icon" :icon-aurora="tab.config.icon" :text="tab.config.title"></f7-link>
     </f7-toolbar>
 
-    <f7-tabs v-if="page && pageType === 'tabs'" :class="{notready: !ready}">
+    <f7-tabs v-if="page && pageType === 'tabs' && visibleToCurrentUser" :class="{notready: !ready}">
       <f7-tab v-for="(tab, idx) in page.slots.default" :key="idx" :tab-active="currentTab === idx">
         <component v-if="currentTab === idx" :is="tabComponent(tab)" :context="tabContext(tab)" @command="onCommand" />
       </f7-tab>
     </f7-tabs>
 
-    <component :is="page.component" v-if="page" :context="context" :class="{notready: !ready}" @command="onCommand" />
+    <component :is="page.component" v-if="page && visibleToCurrentUser" :context="context" :class="{notready: !ready}" @command="onCommand" />
+
+    <empty-state-placeholder v-if="!visibleToCurrentUser" icon="multiply_circle_fill" title="page.unavailable.title" text="page.unavailable.text" />
 
   </f7-page>
 </template>
@@ -82,6 +84,14 @@ export default {
     },
     isAdmin () {
       return this.ready && this.$store.getters.isAdmin
+    },
+    visibleToCurrentUser () {
+      if (!this.page || !this.page.config || !this.page.config.visibleTo) return true
+      const user = this.$store.getters.user
+      if (!user) return false
+      if (user.roles && user.roles.some(r => this.page.config.visibleTo.indexOf('role:' + r) >= 0)) return true
+      if (this.page.config.visibleTo.indexOf('user:' + user.name) >= 0) return true
+      return false
     },
     showBackButton () {
       return this.deep && (!this.page || !this.page.config.sidebar)
