+++ bundles/org.openhab.ui/web/src/pages/settings/pages/pagedesigner-mixin.js
@@ -1,0 +1,260 @@
+import YAML from 'yaml'
+
+export default {
+  data () {
+    return {
+      pageReady: false,
+      loading: false,
+      pageKey: this.$f7.utils.id(),
+      pageYaml: null,
+      previewMode: false,
+      currentTab: 'design',
+      clipboard: null,
+      clipboardType: null,
+      currentComponent: null,
+      currentComponentConfig: null,
+      currentWidget: null,
+      widgetConfigOpened: false,
+      widgetCodeOpened: false,
+      widgetYaml: null
+    }
+  },
+  computed: {
+    ready () {
+      return this.pageReady && this.$store.state.components.widgets != null
+    },
+    context () {
+      return {
+        component: this.page,
+        store: this.$store.getters.trackedItems,
+        editmode: (!this.previewMode) ? {
+          addWidget: this.addWidget,
+          configureWidget: this.configureWidget,
+          configureSlot: this.configureSlot,
+          editWidgetCode: this.editWidgetCode,
+          cutWidget: this.cutWidget,
+          copyWidget: this.copyWidget,
+          pasteWidget: this.pasteWidget,
+          moveWidgetUp: this.moveWidgetUp,
+          moveWidgetDown: this.moveWidgetDown,
+          removeWidget: this.removeWidget
+        } : null,
+        clipboardtype: this.clipboardType
+      }
+    },
+    yamlError () {
+      if (this.currentTab !== 'code') return null
+      try {
+        YAML.parse(this.pageYaml, { prettyErrors: true })
+        return 'OK'
+      } catch (e) {
+        return e
+      }
+    },
+    widgetYamlError () {
+      if (!this.widgetCodeOpened) return null
+      try {
+        YAML.parse(this.widgetYaml, { prettyErrors: true })
+        return 'OK'
+      } catch (e) {
+        return e
+      }
+    }
+  },
+  methods: {
+    onPageAfterIn () {
+      if (window) {
+        window.addEventListener('keydown', this.keyDown)
+      }
+      this.$store.dispatch('startTrackingStates')
+      this.load()
+    },
+    onPageBeforeOut () {
+      if (window) {
+        window.removeEventListener('keydown', this.keyDown)
+      }
+      this.$store.dispatch('stopTrackingStates')
+    },
+    keyDown (ev) {
+      if (ev.ctrlKey || ev.metakKey) {
+        switch (ev.keyCode) {
+          case 82:
+            this.previewMode = !this.previewMode
+            ev.stopPropagation()
+            ev.preventDefault()
+            break
+          case 83:
+            this.save(!this.createMode)
+            ev.stopPropagation()
+            ev.preventDefault()
+            break
+        }
+      }
+    },
+    load () {
+      if (this.loading) return
+      this.loading = true
+
+      if (this.createMode) {
+        this.loading = false
+        this.pageReady = true
+      } else {
+        this.$oh.api.get('/rest/ui/components/ui:page/' + this.uid).then((data) => {
+          this.$set(this, 'page', data)
+          this.pageReady = true
+          this.loading = false
+        })
+      }
+    },
+    save (stay) {
+      if (this.currentTab === 'code' && !this.fromYaml()) return
+      if (!this.page.uid) {
+        this.$f7.dialog.alert('Please give an ID to the page')
+        return
+      }
+      if (!this.page.config.label) {
+        this.$f7.dialog.alert('Please give a label to the page')
+        return
+      }
+      if (!this.createMode && this.uid !== this.page.uid) {
+        this.$f7.dialog.alert('You cannot change the ID of an existing page. Duplicate it with the new ID then delete this one.')
+        return
+      }
+
+      const promise = (this.createMode)
+        ? this.$oh.api.postPlain('/rest/ui/components/ui:page', JSON.stringify(this.page), 'text/plain', 'application/json')
+        : this.$oh.api.put('/rest/ui/components/ui:page/' + this.page.uid, this.page)
+      promise.then((data) => {
+        if (this.createMode) {
+          this.$f7.toast.create({
+            text: 'Page created',
+            destroyOnClose: true,
+            closeTimeout: 2000
+          }).open()
+          this.load()
+        } else {
+          this.$f7.toast.create({
+            text: 'Page updated',
+            destroyOnClose: true,
+            closeTimeout: 2000
+          }).open()
+        }
+        this.$f7.emit('sidebarRefresh', null)
+        if (!stay) this.$f7router.back()
+      }).catch((err) => {
+        this.$f7.toast.create({
+          text: 'Error while saving page: ' + err,
+          destroyOnClose: true,
+          closeTimeout: 2000
+        }).open()
+      })
+    },
+    widgetConfigClosed () {
+      this.currentComponent = null
+      this.currentWidget = null
+      this.widgetConfigOpened = false
+    },
+    updateWidgetConfig () {
+      this.$set(this.currentComponent, 'config', this.currentComponentConfig)
+      this.forceUpdate()
+      this.widgetConfigClosed()
+    },
+    widgetCodeClosed () {
+      this.currentComponent = null
+      this.currentWidget = null
+      this.widgetCodeOpened = false
+    },
+    updateWidgetCode () {
+      const updatedWidget = YAML.parse(this.widgetYaml)
+      this.$set(this.currentComponent, 'config', updatedWidget.config)
+      this.$set(this.currentComponent, 'slots', updatedWidget.slots)
+      this.forceUpdate()
+      this.widgetCodeClosed()
+    },
+    configureWidget (component, parentContext, forceComponentType) {
+      const componentType = forceComponentType || component.component
+      this.currentComponent = null
+      this.currentWidget = null
+      let widgetDefinition
+      if (componentType.indexOf('widget:') === 0) {
+        this.currentWidget = this.$store.getters.widget(componentType.substring(7))
+      } else {
+        // getWidgetDefinition should be defined locally in the page designers SFCs
+        widgetDefinition = this.getWidgetDefinition(componentType)
+        if (!widgetDefinition) {
+          console.warn('Widget not found: ' + componentType)
+          this.$f7.toast.create({
+            text: `This type of component cannot be configured: ${componentType}.`,
+            destroyOnClose: true,
+            closeTimeout: 3000,
+            closeButton: true,
+            closeButtonText: 'Edit YAML',
+            on: {
+              closeButtonClick: () => {
+                this.editWidgetCode(component, parentContext)
+              }
+            }
+          }).open()
+          return
+        }
+        this.currentWidget = widgetDefinition
+      }
+      this.currentComponent = component
+      this.currentComponentConfig = JSON.parse(JSON.stringify(this.currentComponent.config))
+      this.widgetConfigOpened = true
+    },
+    editWidgetCode (component, parentContext, slot) {
+      if (slot && !component.slots) component.slots = {}
+      if (slot && !component.slots[slot]) component.slots[slot] = []
+      this.currentComponent = component
+      this.widgetYaml = YAML.stringify(component)
+      this.widgetCodeOpened = true
+    },
+    cutWidget (component, parentContext, slot = 'default') {
+      this.copyWidget(component, parentContext)
+      this.removeWidget(component, parentContext)
+    },
+    copyWidget (component, parentContext, slot = 'default') {
+      let newClipboard = JSON.stringify(component)
+      this.$set(this, 'clipboard', newClipboard)
+      this.clipboardType = component.component
+    },
+    pasteWidget (component, parentContext, slot = 'default') {
+      if (!this.clipboard) return
+      component.slots[slot].push(JSON.parse(this.clipboard))
+      this.forceUpdate()
+    },
+    moveWidgetUp (component, parentContext, slot = 'default') {
+      let siblings = parentContext.component.slots[slot]
+      let pos = siblings.indexOf(component)
+      if (pos <= 0) return
+      siblings.splice(pos, 1)
+      siblings.splice(pos - 1, 0, component)
+      this.forceUpdate()
+    },
+    moveWidgetDown (component, parentContext, slot = 'default') {
+      let siblings = parentContext.component.slots[slot]
+      let pos = siblings.indexOf(component)
+      if (pos >= siblings.length - 1) return
+      siblings.splice(pos, 1)
+      siblings.splice(pos + 1, 0, component)
+      this.forceUpdate()
+    },
+    removeWidget (component, parentContext, slot = 'default') {
+      parentContext.component.slots[slot].splice(parentContext.component.slots[slot].indexOf(component), 1)
+      this.forceUpdate()
+    },
+    forceUpdate () {
+      this.pageKey = this.$f7.utils.id()
+    },
+    togglePreviewMode (value) {
+      if (value === undefined) value = !this.previewMode
+      if (value === true) {
+        if (this.currentTab === 'code') {
+          if (!this.fromYaml()) return
+        }
+      }
+      this.previewMode = value
+    }
+  }
+}
